Updating product brand Issue fixed

+++ src/modules/admin/features/attributes/api/attributes.api.js
@@ -4,7 +4,7 @@ import { handleApiError } from "../../../../../utils/apiError";
 // Create a new attribute
 export const createAttribute = async (attributeData) => {
      try{
-        const res = await API.post("/attributes", attributeData);
+        const res = await API.post("/attributes/admin", attributeData);
         return res.data;
      } catch (error) {
         throw handleApiError(error, "Failed to create attribute");
@@ -14,7 +14,7 @@ export const createAttribute = async (attributeData) => {
 // Create a new value for a specific attribute
 export const createAttributeValue = async (attributeId, valueData) => {
   try {
-    const res = await API.post(`/attributes/${attributeId}/value`, valueData);
+    const res = await API.post(`/attributes/admin/${attributeId}/value`, valueData);
     return res.data;
   } catch (err) {
     throw handleApiError(err, "Failed to create attribute value");
@@ -24,7 +24,7 @@ export const createAttributeValue = async (attributeId, valueData) => {
 // get all attributes with their values devided into categories for attribute page 
 export const fetchAttributesCatalog = async () => {
   try {
-    const res = await API.get("/attributes");
+    const res = await API.get("/attributes/admin");
     return res.data;
   } catch (error) {
     throw handleApiError(error, "Failed to fetch attributes catalog");
@@ -34,7 +34,7 @@ export const fetchAttributesCatalog = async () => {
 // for delete attribute in attribute page
 export const deleteAttribute = async(attributeId) => {
   try{
-    const res = await API.delete(`/attributes/${attributeId}`);;
+    const res = await API.delete(`/attributes/admin/${attributeId}`);;
     return res.data;
   } catch (err) {
     throw handleApiError(err, "Failed to delete attribute");
@@ -44,7 +44,7 @@ export const deleteAttribute = async(attributeId) => {
 // for delete attribute value in attribute page
 export const deleteAttributeValue = async (attributeId, attributeValueId) => {
   try {
-    const res = await API.delete(`/attributes/${attributeId}/value/${attributeValueId}`);
+    const res = await API.delete(`/attributes/admin/${attributeId}/value/${attributeValueId}`);
     return res.data;
   } catch (err) {
     throw handleApiError(err, "Failed to delete attribute value");
